Reverted to edit edt-e4c85821-fe1a-443a-93c2-3bcd625d2743: "Set white background for header
The BookPreviewHeader component's background was changed to solid white."
