+++ src/components/book-progress/BookEditorModals.tsx
@@ -1,9 +1,7 @@
 
-import { useState } from "react";
 import CoverEditor from "../cover/CoverEditor";
 import BookPreview from "../book/BookPreview";
 import { CoverData } from "../cover/CoverTypes";
-import { useToast } from "@/hooks/use-toast";
 
 interface BookEditorModalsProps {
   profileId: string;
@@ -12,7 +10,7 @@ interface BookEditorModalsProps {
   coverData: CoverData;
   onCloseCoverEditor: () => void;
   onClosePreview: () => void;
-  onSaveCover: (coverData: CoverData) => Promise<boolean>;
+  onSaveCover: (coverData: CoverData) => Promise<void>;
 }
 
 const BookEditorModals = ({
@@ -24,39 +22,13 @@ const BookEditorModals = ({
   onClosePreview,
   onSaveCover
 }: BookEditorModalsProps) => {
-  const { toast } = useToast();
-  const [isSaving, setIsSaving] = useState(false);
-
-  const handleSaveCover = async (data: CoverData) => {
-    setIsSaving(true);
-    try {
-      const success = await onSaveCover(data);
-      if (!success) {
-        toast({
-          variant: "destructive",
-          title: "Error",
-          description: "Failed to save cover data. Please try again."
-        });
-      }
-    } catch (error) {
-      console.error("Error in handleSaveCover:", error);
-      toast({
-        variant: "destructive",
-        title: "Error",
-        description: "Failed to save cover data. Please try again."
-      });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
   return (
     <>
       <CoverEditor
         profileId={profileId}
         open={isEditorOpen}
         onClose={onCloseCoverEditor}
-        onSave={handleSaveCover}
+        onSave={onSaveCover}
         initialCoverData={coverData}
       />
