+++ src/components/cover/editor/EditorControlPanel.tsx
@@ -19,7 +19,6 @@ interface EditorControlPanelProps {
   onTextColorChange: (color: string, type: 'title' | 'author') => void;
   onFontSizeChange: (value: number[], type: 'title' | 'author') => void;
   onLayoutChange: (layout: 'centered' | 'top' | 'bottom') => void;
-  isSaving?: boolean;
 }
 
 const EditorControlPanel = ({
@@ -34,7 +33,6 @@ const EditorControlPanel = ({
   onTextColorChange,
   onFontSizeChange,
   onLayoutChange,
-  isSaving = false,
 }: EditorControlPanelProps) => {
   const [activeTab, setActiveTab] = useState("background");
 
@@ -77,14 +75,11 @@ const EditorControlPanel = ({
       </Tabs>
       
       <div className="absolute bottom-4 left-4 right-4 flex justify-end gap-2">
-        <Button variant="outline" onClick={onCancel} disabled={isSaving || isUploading}>
+        <Button variant="outline" onClick={onCancel}>
           Cancel
         </Button>
-        <Button 
-          onClick={onSave} 
-          disabled={isSaving || isUploading}
-        >
-          {isSaving ? 'Saving...' : 'Save Cover'}
+        <Button onClick={onSave}>
+          Save Cover
         </Button>
       </div>
     </div>
