+++ src/hooks/useCoverData.ts
@@ -1,26 +1,18 @@
-
 import { useState, useEffect, useCallback } from "react";
+import { supabase } from "@/integrations/supabase/client";
 import { useToast } from "@/hooks/use-toast";
 import { CoverData, DEFAULT_COVER_DATA } from "@/components/cover/CoverTypes";
 import { useAuth } from "@/contexts/AuthContext";
-import { fetchCoverData, saveCoverData as apiSaveCoverData } from "@/api/coverDataService";
-import { 
-  getCachedCoverData, 
-  cacheCoverData, 
-  markCoverDataSaved, 
-  markCoverDataSaving 
-} from "@/utils/coverDataCache";
 
 export function useCoverData(profileId: string) {
   const [coverData, setCoverData] = useState<CoverData | null>(null);
   const [isLoading, setIsLoading] = useState(true);
-  const [isSaving, setIsSaving] = useState(false);
   const [error, setError] = useState<Error | null>(null);
   const { toast } = useToast();
   const { isAuthenticated } = useAuth();
 
   // Function to fetch cover data
-  const refreshCoverData = useCallback(async () => {
+  const fetchCoverData = useCallback(async () => {
     if (!profileId) {
       setIsLoading(false);
       return;
@@ -31,15 +23,27 @@ export function useCoverData(profileId: string) {
       console.log('Fetching cover data for profile:', profileId);
       
       // First, try to get data from localStorage cache for immediate display
-      const cachedData = getCachedCoverData(profileId);
+      const cachedData = localStorage.getItem(`cover_data_${profileId}`);
       if (cachedData) {
-        console.log('Using cached cover data while fetching from server:', cachedData);
-        setCoverData(cachedData);
+        const parsedData = JSON.parse(cachedData) as CoverData;
+        console.log('Using cached cover data while fetching from server:', parsedData);
+        setCoverData(parsedData);
       }
       
       // Then fetch from server to ensure we have the latest
-      const data = await fetchCoverData(profileId);
+      const { data, error } = await supabase
+        .from('book_covers')
+        .select('cover_data, updated_at')
+        .eq('profile_id', profileId)
+        .maybeSingle();
 
+      if (error) {
+        console.error('Error fetching cover data:', error);
+        throw error;
+      }
+
+      console.log('Received cover data from database:', data);
+      
       if (data && data.cover_data) {
         // Explicitly cast and set the cover data
         const typedCoverData = data.cover_data as CoverData;
@@ -47,8 +51,9 @@ export function useCoverData(profileId: string) {
         setCoverData(typedCoverData);
         
         // Update local cache
-        cacheCoverData(profileId, typedCoverData);
-        markCoverDataSaved(profileId, data.updated_at);
+        localStorage.setItem(`cover_data_${profileId}`, JSON.stringify(typedCoverData));
+        localStorage.setItem(`cover_data_last_fetched_${profileId}`, new Date().toISOString());
+        localStorage.setItem(`cover_data_updated_at_${profileId}`, data.updated_at);
       } else {
         // If no cover data exists yet, use defaults but don't save until user makes changes
         console.log('No cover data found, using defaults');
@@ -66,7 +71,7 @@ export function useCoverData(profileId: string) {
         description: "Failed to load cover data",
       });
       // Even on error, use the default data if we don't have cached data
-      if (!getCachedCoverData(profileId)) {
+      if (!localStorage.getItem(`cover_data_${profileId}`)) {
         setCoverData(DEFAULT_COVER_DATA);
       }
     } finally {
@@ -78,15 +83,15 @@ export function useCoverData(profileId: string) {
   useEffect(() => {
     console.log('Profile ID in useCoverData:', profileId);
     if (profileId) {
-      refreshCoverData();
+      fetchCoverData();
     } else {
       // If no profileId, set defaults and not loading
       setCoverData(DEFAULT_COVER_DATA);
       setIsLoading(false);
     }
-  }, [profileId, refreshCoverData]);
+  }, [profileId, fetchCoverData]);
 
-  const saveCoverData = async (newCoverData: CoverData): Promise<boolean> => {
+  const saveCoverData = async (newCoverData: CoverData) => {
     // Check for profile ID first
     if (!profileId) {
       console.error("Cannot save: No profile ID provided");
@@ -100,19 +105,31 @@ export function useCoverData(profileId: string) {
 
     try {
       console.log('Saving cover data:', newCoverData);
-      setIsSaving(true);
       
       // Update local cache immediately for responsiveness
-      cacheCoverData(profileId, newCoverData);
-      markCoverDataSaving(profileId);
+      localStorage.setItem(`cover_data_${profileId}`, JSON.stringify(newCoverData));
+      localStorage.setItem(`cover_data_saving_${profileId}`, new Date().toISOString());
       
       // Update local state right away
       setCoverData(newCoverData);
       
-      // Direct call to the Supabase Edge Function via our API service
-      await apiSaveCoverData(profileId, newCoverData);
+      // Direct call to the Supabase Edge Function
+      const { data, error } = await supabase.functions.invoke('save-cover-data', {
+        method: 'POST',
+        body: {
+          profileId: profileId,
+          coverData: newCoverData
+        },
+      });
+
+      if (error) {
+        console.error('Error from save-cover-data function:', error);
+        throw new Error(error.message || 'Failed to save cover data');
+      }
       
-      markCoverDataSaved(profileId);
+      console.log('Response from save-cover-data function:', data);
+      
+      localStorage.setItem(`cover_data_saved_at_${profileId}`, new Date().toISOString());
       
       toast({
         title: "Cover saved",
@@ -128,17 +145,14 @@ export function useCoverData(profileId: string) {
         description: "Failed to save cover data",
       });
       return false;
-    } finally {
-      setIsSaving(false);
     }
   };
 
   return {
     coverData: coverData || DEFAULT_COVER_DATA,
     isLoading,
-    isSaving,
     error,
     saveCoverData,
-    refreshCoverData
+    refreshCoverData: fetchCoverData
   };
 }
